id is now consistently one variable type

+++ src/main/java/TechTigers/BicycleBuddies/data/RideRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RideRepository extends JpaRepository<Ride, Long> {
+public interface RideRepository extends JpaRepository<Ride, Integer> {
     // JpaRepository already provides basic CRUD operations, so no need for additional methods.
 }
